Factor out FakeURLSession / FakeURLDataTask to their own helper files
