+++ Tests/TezosKit/AbstractClientTest.swift
@@ -3,49 +3,6 @@
 @testable import TezosKit
 import XCTest
 
-/// A fake URLSession that will return data tasks which will call completion handlers with the given parameters.
-public class FakeURLSession: URLSession {
-  public var urlResponse: URLResponse?
-  public var data: Data?
-  public var error: Error?
-
-  public override func dataTask(
-    with request: URLRequest,
-    completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void
-  ) -> URLSessionDataTask {
-    return FakeURLSessionDataTask(
-      urlResponse: urlResponse,
-      data: data,
-      error: error,
-      completionHandler: completionHandler
-    )
-  }
-}
-
-/// A fake data task that will immediately call completion.
-public class FakeURLSessionDataTask: URLSessionDataTask {
-  private let urlResponse: URLResponse?
-  private let data: Data?
-  private let fakedError: Error?
-  private let completionHandler: (Data?, URLResponse?, Error?) -> Void
-
-  public init(
-    urlResponse: URLResponse?,
-    data: Data?,
-    error: Error?,
-    completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void
-    ) {
-    self.urlResponse = urlResponse
-    self.data = data
-    self.fakedError = error
-    self.completionHandler = completionHandler
-  }
-
-  public override func resume() {
-    completionHandler(data, urlResponse, fakedError)
-  }
-}
-
 class AbstractClientTest: XCTestCase {
   public var abstractClient: AbstractClient?
